requirements.txt
asgiref==3.5.0
autopep8==1.6.0
boto3==1.20.54
botocore==1.23.54
certifi==2021.10.8
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-heroku==0.3.1
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
filelock==3.4.2
gunicorn==20.1.0
jmespath==0.10.0
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.4.1
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
whitenoise==6.0.0
